refactored extent functions and added extent to dataset table flow is not working
